rename ImageStatus to TrackImage
