+++ src/js/TrackImage.js
